[Bugfix] Move foreign keys to applicant model (#47)
* Moved foreign keys to applicant model, modularized seeding script, fixed broken APIs

* Update GraphQL types

* Added temporary fix to prop types of Request Info page

* Remove comment
