+++ .prettierignore
@@ -6,3 +6,4 @@ coverage
 
 # Ignore autogenerated files:
 lib/graphql/types.ts
+**/*.prisma
