+++ prisma/schema.prisma
@@ -8,29 +8,31 @@ datasource db {
 }
 
 model Applicant {
-  id                 Int                 @id @default(autoincrement())
-  firstName          String              @map("first_name") @db.VarChar(255)
-  middleName         String?             @map("middle_name") @db.VarChar(255)
-  lastName           String              @map("last_name") @db.VarChar(255)
-  dateOfBirth        DateTime            @map("date_of_birth") @db.Date
-  gender             Gender
-  customGender       String?             @map("custom_gender") @db.VarChar(255)
-  email              String?             @unique @db.VarChar(255)
-  phone              String              @db.VarChar(50)
-  province           Province
-  city               String              @db.VarChar(255)
-  addressLine1       String              @map("address_line_1") @db.VarChar(255)
-  addressLine2       String?             @map("address_line_2") @db.VarChar(255)
-  postalCode         String              @map("postal_code") @db.Char(6)
-  rcdUserId          Int?                @unique @map("rcd_user_id")
-  status             ApplicantStatus?
-  acceptedTOC        DateTime?           @map("accepted_toc") @db.Timestamptz(6)
-  createdAt          DateTime            @default(now()) @map("created_at") @db.Timestamptz(6)
-  updatedAt          DateTime            @default(now()) @updatedAt @map("updated_at") @db.Timestamptz(6)
-  applications       Application[]
-  guardian           Guardian?
-  medicalInformation MedicalInformation?
-  permits            Permit[]
+  id                   Int                @id @default(autoincrement())
+  firstName            String             @map("first_name") @db.VarChar(255)
+  middleName           String?            @map("middle_name") @db.VarChar(255)
+  lastName             String             @map("last_name") @db.VarChar(255)
+  dateOfBirth          DateTime           @map("date_of_birth") @db.Date
+  gender               Gender
+  customGender         String?            @map("custom_gender") @db.VarChar(255)
+  email                String?            @unique @db.VarChar(255)
+  phone                String             @db.VarChar(50)
+  province             Province
+  city                 String             @db.VarChar(255)
+  addressLine1         String             @map("address_line_1") @db.VarChar(255)
+  addressLine2         String?            @map("address_line_2") @db.VarChar(255)
+  postalCode           String             @map("postal_code") @db.Char(6)
+  rcdUserId            Int?               @unique @map("rcd_user_id")
+  status               ApplicantStatus?
+  acceptedTOC          DateTime?          @map("accepted_toc") @db.Timestamptz(6)
+  guardianId           Int                @unique @map("guardian_id")
+  medicalInformationId Int                @unique @map("medical_information_id")
+  createdAt            DateTime           @default(now()) @map("created_at") @db.Timestamptz(6)
+  updatedAt            DateTime           @default(now()) @updatedAt @map("updated_at") @db.Timestamptz(6)
+  guardian             Guardian           @relation("applicantsToguardians", fields: [guardianId], references: [id])
+  medicalInformation   MedicalInformation @relation("applicantsTomedical_information", fields: [medicalInformationId], references: [id])
+  applications         Application[]
+  permits              Permit[]
 
   @@map("applicants")
 }
@@ -119,41 +121,39 @@ model Employee {
 }
 
 model Guardian {
-  id           Int       @id @default(autoincrement())
-  firstName    String    @map("first_name") @db.VarChar(255)
-  middleName   String?   @map("middle_name") @db.VarChar(255)
-  lastName     String    @map("last_name") @db.VarChar(255)
-  addressLine1 String    @map("address_line_1") @db.VarChar(255)
-  addressLine2 String?   @map("address_line_2") @db.VarChar(255)
-  city         String    @db.VarChar(255)
+  id           Int        @id @default(autoincrement())
+  firstName    String     @map("first_name") @db.VarChar(255)
+  middleName   String?    @map("middle_name") @db.VarChar(255)
+  lastName     String     @map("last_name") @db.VarChar(255)
+  addressLine1 String     @map("address_line_1") @db.VarChar(255)
+  addressLine2 String?    @map("address_line_2") @db.VarChar(255)
+  city         String     @db.VarChar(255)
   province     Province
-  postalCode   String    @map("postal_code") @db.Char(6)
-  phone        String    @db.VarChar(50)
-  relationship String    @db.VarChar(255)
+  postalCode   String     @map("postal_code") @db.Char(6)
+  phone        String     @db.VarChar(50)
+  relationship String     @db.VarChar(255)
   notes        String?
-  applicantId  Int       @unique @map("applicant_id")
-  createdAt    DateTime  @default(now()) @map("created_at") @db.Timestamptz(6)
-  updatedAt    DateTime  @default(now()) @updatedAt @map("updated_at") @db.Timestamptz(6)
-  applicant    Applicant @relation(fields: [applicantId], references: [id])
+  createdAt    DateTime   @default(now()) @map("created_at") @db.Timestamptz(6)
+  updatedAt    DateTime   @default(now()) @updatedAt @map("updated_at") @db.Timestamptz(6)
+  applicant    Applicant? @relation("applicantsToguardians")
 
   @@map("guardians")
 }
 
 model MedicalInformation {
-  id                  Int       @id @default(autoincrement())
-  disability          String    @db.VarChar(255)
-  affectsMobility     Boolean   @default(false) @map("affects_mobility")
-  mobilityAidRequired Boolean   @default(false) @map("mobility_aid_required")
-  cannotWalk100m      Boolean   @default(false) @map("cannot_walk_100m")
+  id                  Int        @id @default(autoincrement())
+  disability          String     @db.VarChar(255)
+  affectsMobility     Boolean    @default(false) @map("affects_mobility")
+  mobilityAidRequired Boolean    @default(false) @map("mobility_aid_required")
+  cannotWalk100m      Boolean    @default(false) @map("cannot_walk_100m")
   notes               String?
-  certificationDate   DateTime? @map("certification_date") @db.Date
+  certificationDate   DateTime?  @map("certification_date") @db.Date
   aid                 Aid[]
-  physicianId         Int       @map("physician_id")
-  applicantId         Int       @unique @map("applicant_id")
-  createdAt           DateTime  @default(now()) @map("created_at") @db.Timestamptz(6)
-  updatedAt           DateTime  @default(now()) @map("updated_at") @db.Timestamptz(6)
-  applicant           Applicant @relation(fields: [applicantId], references: [id])
-  physician           Physician @relation(fields: [physicianId], references: [id])
+  physicianId         Int        @map("physician_id")
+  createdAt           DateTime   @default(now()) @map("created_at") @db.Timestamptz(6)
+  updatedAt           DateTime   @default(now()) @map("updated_at") @db.Timestamptz(6)
+  physician           Physician  @relation(fields: [physicianId], references: [id])
+  applicant           Applicant? @relation("applicantsTomedical_information")
 
   @@map("medical_information")
 }
